requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.32
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==43.0.1
Werkzeug==3.0.1
